requirements.txt
click
ffmpeg-python
opencv-python
numpy
tqdm
pillow
customtkinter
tkinterdnd2
scenedetect
setuptools